make DateUtils.swift for utility functions about date things
