+++ BusTimerJSON/ViewController.swift
@@ -7,84 +7,6 @@
 import UIKit
 import SwiftyJSON
 
-//return date time info as string (yyyy/M/dd H:mm)
-func getDateTime() -> String{
-    let date = Date()
-    let dateFormatter = DateFormatter()
-    dateFormatter.dateStyle = .medium
-    dateFormatter.timeStyle = .none
-    dateFormatter.locale = Locale(identifier: "ja_JP")
-    dateFormatter.timeZone = TimeZone(abbreviation: "JST")
-    let today = dateFormatter.string(from: date)
-    //    print("getDateTime:",today)
-    return today //return 2019/01/05
-}
-
-func identifyDayType(today:String, holidays:[String]) -> String{
-    let dayOfWeek = getDayOfWeek()
-    if(isHoliday(today: today, holidays: holidays)){
-        print("Holiday スケジュール")
-        //        return "holiday"
-        return "sun"
-    } else if(dayOfWeek == 1){
-        print("Sunday スケジュール")
-        return "sun"
-    } else if(dayOfWeek == 7){
-        print("Saturday スケジュール")
-        return "sat"
-        //        return "saturday"
-    } else {
-        print("Weekday スケジュール")
-        return "weekday"
-    }
-}
-
-func isHoliday(today:String, holidays:[String]) -> Bool {
-    var result = false
-    for i in holidays {
-        if(today == i){
-            result = true
-        }
-    }
-    return result
-}
-
-func getDayOfWeek() -> Int{
-    let date = Date()
-    let calendar = Calendar.current
-    let dayOfWeek = calendar.component(.weekday, from: date)
-    //    print("Weekday \(weekday)") // 1, 2, 3, .... 2 is Monday
-    return dayOfWeek
-}
-
-func strToDate(str:String)-> Date?{
-    //    print("str",str)
-    let dateFormatter = DateFormatter()
-    dateFormatter.dateFormat = "yyyy/MM/dd H:mm"
-    //    dateFormatter.locale = Locale(identifier: "en_US_POSIX") //王道パターン
-    let date = dateFormatter.date(from: str)
-    //    print(dateFormatter.string(from: date!)) //値自体は変わらない
-    return date
-}
-
-func dateToStr(dateObj: Date?) -> String{
-    let dateFormatter = DateFormatter()
-    dateFormatter.dateFormat = "yyyy/MM/dd H:mm"
-    return dateFormatter.string(from: dateObj!)
-}
-
-func jsonToDateObj(jsonObj: JSON?) -> Date?{
-    let dateFormatter = DateFormatter()
-    dateFormatter.dateFormat = "yyyy/MM/dd"
-    // ex: 2019/01/11
-    let currDateOnly = dateFormatter.string(from: currUserTime)
-    // "currDateOnly + busHour:busMin"   ex: "2019/01/11 09:25"
-    let jsonTimeStr = "\(currDateOnly) \(jsonObj!["hour"].intValue):\(jsonObj!["min"].intValue)"
-    // convert string to dateTime object
-    let busDateObj = strToDate(str:jsonTimeStr)
-    return busDateObj
-}
-
 func getNextBus(callback: @escaping (Date?)->()) {
     // User inputs
     
@@ -95,11 +17,11 @@ func getNextBus(callback: @escaping (Date?)->()) {
     
     let arrayKeys = Array(holidaysJson.dictionaryValue.keys)
     // get today in a format comparable
-    var todayFormatted = getDateTime()
+    var todayFormatted = DateUtils.getDateTime()
     print(todayFormatted)
     todayFormatted = todayFormatted.replacingOccurrences(of: "/", with: "-", options: .literal, range: nil)
     // Identify if today is weekday, saturday or holiday
-    userWeek = identifyDayType(today: todayFormatted, holidays: arrayKeys)
+    userWeek = DateUtils.identifyDayType(today: todayFormatted, holidays: arrayKeys)
     
     
     // Get next bus information
@@ -108,7 +30,7 @@ func getNextBus(callback: @escaping (Date?)->()) {
     // check if the last bus has left
     let lastBus = busSchedule.last
     // convert JSON to date object
-    let lastBusObj = jsonToDateObj(jsonObj: lastBus)
+    let lastBusObj = DateUtils.jsonToDateObj(jsonObj: lastBus)
     // Compare
     if lastBusObj! < currUserTime{
         print("Last bus has left!")
@@ -117,10 +39,10 @@ func getNextBus(callback: @escaping (Date?)->()) {
         // Find the next bus
         for bus in busSchedule{
             // convert JSON to date object
-            let busTimeObj = jsonToDateObj(jsonObj: bus)
+            let busTimeObj = DateUtils.jsonToDateObj(jsonObj: bus)
             // Compare
             if busTimeObj! > currUserTime{
-                print("Found next bus at",dateToStr(dateObj: busTimeObj))
+                print("Found next bus at", DateUtils.dateToStr(dateObj: busTimeObj))
                 _nextBus = busTimeObj
                 break
             }
@@ -191,7 +113,7 @@ class ViewController: UIViewController {
                 //show next bus
                 DispatchQueue.main.async {
                     self.busInfo.numberOfLines = 0
-                    self.busInfo.text = dateToStr(dateObj: nextBusDateObj)
+                    self.busInfo.text = DateUtils.dateToStr(dateObj: nextBusDateObj)
                     nextBusTime = nextBusDateObj
                 }
             }
